requirements.txt
# 核心依赖
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tkcalendar>=1.6.0,<2.0.0

# 测试依赖
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-xdist>=3.0.0,<4.0.0
pytest-timeout>=2.1.0,<3.0.0
pytest-html>=3.1.0,<4.0.0

# 代码质量工具
black>=22.0.0,<24.0.0
flake8>=5.0.0,<7.0.0
mypy>=1.0.0,<2.0.0
isort>=5.12.0,<6.0.0

# 打包依赖
pyinstaller>=6.0.0

# 虚拟环境管理
virtualenv>=20.0.0,<21.0.0